+++ src/index.jsx
@@ -1,39 +1,9 @@
 /* eslint-disable linebreak-style */
-import React, { useState } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
 
 // 컴포넌트
-import Counter from './Counter';
-import NumberButtonList from './NumberButtonList';
-import ResetButton from './ResetButton';
-
-function App() {
-  const [count, setCount] = useState(0);
-
-  function handleClickCounter() {
-    setCount(count + 1);
-  }
-
-  function handleClickNumberButton(number = 0) {
-    setCount(count + number);
-  }
-
-  function handleClickResetButton() {
-    setCount(0);
-  }
-
-  return (
-    <div className="container">
-      <h1>Counter</h1>
-
-      <Counter count={count} onClick={handleClickCounter} />
-
-      <NumberButtonList onClick={handleClickNumberButton} />
-
-      <ResetButton onClick={handleClickResetButton} />
-    </div>
-  );
-}
+import App from './App';
 
 ReactDOM.render(
   <App />,
